+++ lib/src/builder/run_builder.dart
@@ -37,7 +37,7 @@ class RunCMakeBuilder {
   final BuildMode buildMode;
 
   /// -G
-  final Generator generator;
+  Generator generator;
 
   /// -T
   final String? toolset;
@@ -230,6 +230,8 @@ class RunCMakeBuilder {
     final defs = <String>[];
     final toolchain = await androidToolchainCmake();
     defs.add('-DCMAKE_TOOLCHAIN_FILE=${toolchain.normalizePath().toFilePath()}');
+
+    generator == Generator.defaultGenerator ? generator = Generator.ninja : generator = generator;
 
     // The Android Gradle plugin does not honor API level 19 and 20 when
     // invoking clang. Mimic that behavior here.